refactor: separate interface dm
