+++ frontend/src/types/interface/dm.interface.ts
@@ -1,0 +1,7 @@
+
+interface Message {
+  time: string | Date;
+  sender: number; // mystate id
+  receiver: string; // receiver nickname
+  content: string; //
+}
